refactor: move SearchDropdown
