+++ src/components/TeacherComponents/SearchDropdown/SearchDropdown.tsx
@@ -1,7 +1,6 @@
 import { FC, useState } from "react";
 
-import { SearchResultsItemProps } from "../TeacherComponents/SearchResultsItem/SearchResultsItem";
-
+import { SearchResultsItemProps } from "@/components/TeacherComponents/SearchResultsItem/SearchResultsItem";
 import MiniDropDown from "@/components/SharedComponents/Button/MiniDropDownButton/MiniDropDown";
 import Flex from "@/components/SharedComponents/Flex";
 import Box from "@/components/SharedComponents/Box";
