+++ src/components/TeacherComponents/SearchDropdown/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./SearchDropdown";
