+++ src/components/TeacherComponents/SearchResultsItem/SearchResultsItem.tsx
@@ -4,7 +4,7 @@ import SearchResultsSubjectIcon from "@/components/TeacherComponents/SearchResul
 import LessonMetadata from "@/components/SharedComponents/LessonMetadata";
 import TagPromotional from "@/components/SharedComponents/TagPromotional";
 import OakLink from "@/components/OakLink";
-import SearchDropdown from "@/components/SearchDropdown/SearchDropdown";
+import SearchDropdown from "@/components/TeacherComponents/SearchDropdown";
 import Icon from "@/components/SharedComponents/Icon";
 import { Heading, P, Span } from "@/components/SharedComponents/Typography";
 import Flex from "@/components/SharedComponents/Flex";
